LayoutTests/svg/dynamic-updates/SVGFEGaussianBlurElement-svgdom-edgeMode-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'edgeMode' property of the SVGFEGaussianBlurElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gaussianBlurElement.edgeMode.baseVal is SVGFEGaussianBlurElement.SVG_EDGEMODE_WRAP
PASS gaussianBlurElement.edgeMode.baseVal is SVGFEGaussianBlurElement.SVG_EDGEMODE_DUPLICATE
PASS successfullyParsed is true

TEST COMPLETE
